requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pypdf==6.0.0
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.9
faiss-cpu==1.12.0

